+++ packages/@expo/cli/src/start/server/metro/TerminalReporter.types.ts
@@ -2,6 +2,8 @@ import type { ReportableEvent } from 'metro';
 import type { TerminalReportableEvent } from 'metro/src/lib/TerminalReporter';
 import type { Terminal } from 'metro-core';
 
+import { MetroEnvironment } from '../middleware/metroOptions';
+
 export type GlobalCacheDisabledReason = 'too_many_errors' | 'too_many_misses';
 
 export type BundleDetails = {
@@ -11,7 +13,7 @@ export type BundleDetails = {
   entryFile: string;
   minify: boolean;
   platform: string | null | undefined;
-  customTransformOptions?: { environment?: 'node' };
+  customTransformOptions?: { environment?: MetroEnvironment };
   runtimeBytecodeVersion: number | null | undefined;
 };