+++ packages/@expo/metro-config/src/serializer/environmentVariableSerializerPlugin.ts
@@ -34,13 +34,14 @@ function isServerEnvironment(graph: ReadOnlyGraph, options: SerializerOptions): 
   if (!graph.transformOptions.customTransformOptions) {
     if (options.sourceUrl) {
       const env = getTransformEnvironment(options.sourceUrl);
-      return env === 'node';
+      return env === 'node' || env === 'react-server';
     }
     return false;
   }
 
   // Other requests will use customTransformOptions.environment.
-  return graph.transformOptions.customTransformOptions.environment === 'node';
+  const env = graph.transformOptions.customTransformOptions.environment;
+  return env === 'node' || env === 'react-server';
 }
 
 /** Strips the process.env polyfill in server environments to allow for accessing environment variables off the global. */